+++ Firefly-iii-pp-Runner/Firefly-pp-Runner/Models/ThunderClient/Collection.cs
@@ -8,7 +8,7 @@ namespace Firefly_iii_pp_Runner.Models.ThunderClient
         public string Id { get; set; } = Guid.NewGuid().ToString();
         public string ColName { get; set; }
         public DateTime Created { get; set; } = DateTime.Now;
-        public int SortNum { get; set; }
+        public double SortNum { get; set; }
         public List<Folder> Folders { get; set; } = new List<Folder>();
         public Settings Settings { get; set; }
     }
